Removed usage of is-boolean-object
This package was only used once, and incorrectly.
is-boolean-object checks whether the input is `true`, `false`
or an object created with the `Boolean` constructor.
I doubt that this last case was really intended and properly supported
since the Typescript type is `boolean` (so, only true/false) and
the only code that uses the boolean is this one:

```ts
if (useWrapWord) {
```

however, `if (new Boolean(false))` yields `true`.