Removed `is-number-object` dependency
This library checks whether the input is a number or a Number object
(i.e. created with `new Number(5)`). I never saw anyone using the latter
syntax, and I'm not sure the 17+kB-and-10-files dependency is
really useful.

This is a breaking change since `new Number(5)` will not be allowed
anymore for column width. Does anyone really do that?

This change is in line with the typescript type annotations for the width,
which only allow `number`s to be provided.