Removed the dependency on `is-string`
The package tests whether the input is a string or a string object
(i.e. created with `new Object('hello')`). I'm not sure we should
support the latter syntax, especially since the string is passed
to string-width, which explicitely returns 0 if `typeof` doesn't
return `string` : https://github.com/sindresorhus/string-width/blob/89eb79aafd2b7db982dc1f5498424232d55dee53/index.js#L6

This is a breaking change in the strict sense, because the behavior
would change if such "string-object" was indeed passed as a value,
but that would result in an empty cell, so that probably doesn't matter
and avoids a dependency on a 16+kB and 10 files